refactor(core): reduce core package size
