+++ packages/core/src/models/Field.ts
@@ -1,10 +1,4 @@
-import {
-  FormPath,
-  isValid,
-  FormPathPattern,
-  isEmpty,
-  toArr,
-} from '@formily/shared'
+import { isValid, isEmpty, toArr, FormPathPattern } from '@formily/shared'
 import {
   ValidatorTriggerType,
   parseValidatorDescriptions,
@@ -17,7 +11,6 @@ import {
   toJS,
   action,
 } from '@formily/reactive'
-import { Form } from './Form'
 import {
   JSXComponent,
   LifeCycleTypes,
@@ -25,11 +18,7 @@ import {
   FeedbackMessage,
   IFieldCaches,
   IFieldRequests,
-  FieldDisplayTypes,
-  FieldPatternTypes,
   FieldValidator,
-  FieldDecorator,
-  FieldComponent,
   FieldDataSource,
   ISearchFeedback,
   IFieldProps,
@@ -39,8 +28,6 @@ import {
   IModelGetter,
 } from '../types'
 import {
-  buildNodeIndexes,
-  initFieldUpdate,
   updateFeedback,
   queryFeedbacks,
   allowAssignDefaultValue,
@@ -62,18 +49,18 @@ import {
   initializeStart,
   initializeEnd,
 } from '../shared/internals'
-import { Query } from './Query'
+import { Form } from './Form'
+import { BaseField } from './BaseField'
 export class Field<
   Decorator extends JSXComponent = any,
   Component extends JSXComponent = any,
   TextType = any,
   ValueType = any
-> {
+> extends BaseField<Decorator, Component, TextType> {
   displayName = 'Field'
-  title: TextType
-  description: TextType
-  selfDisplay: FieldDisplayTypes
-  selfPattern: FieldPatternTypes
+
+  props: IFieldProps<Decorator, Component, TextType, ValueType>
+
   loading: boolean
   validating: boolean
   submitting: boolean
@@ -82,28 +69,11 @@ export class Field<
   visited: boolean
   inputValue: ValueType
   inputValues: any[]
-  initialized: boolean
   dataSource: FieldDataSource
-  mounted: boolean
-  unmounted: boolean
-  content: any
-  data: any
   validator: FieldValidator
-  decoratorType: Decorator
-  decoratorProps: Record<string, any>
-  componentType: Component
-  componentProps: Record<string, any>
   feedbacks: IFieldFeedback[]
-  address: FormPath
-  path: FormPath
-
-  form: Form
-  designable: boolean
-  props: IFieldProps<Decorator, Component, TextType, ValueType>
-
   caches: IFieldCaches = {}
   requests: IFieldRequests = {}
-  disposers: (() => void)[] = []
 
   constructor(
     address: FormPathPattern,
@@ -111,6 +81,7 @@ export class Field<
     form: Form,
     designable: boolean
   ) {
+    super()
     this.form = form
     this.props = props
     this.designable = designable
@@ -123,12 +94,7 @@ export class Field<
     initializeEnd()
   }
 
-  protected makeIndexes(address: FormPathPattern) {
-    this.form.fields[address.toString()] = this
-    buildNodeIndexes(this, address)
-  }
-
-  protected initialize() {
+  initialize() {
     this.initialized = false
     this.loading = false
     this.validating = false
@@ -165,7 +131,7 @@ export class Field<
     this.component = toArr(this.props.component)
   }
 
-  protected makeObservable() {
+  makeObservable() {
     if (this.designable) return
     define(this, {
       title: observable.ref,
@@ -246,7 +212,7 @@ export class Field<
     })
   }
 
-  protected makeReactive() {
+  makeReactive() {
     if (this.designable) return
     this.disposers.push(
       reaction(
@@ -299,38 +265,6 @@ export class Field<
       )
     )
     createReactions(this)
-  }
-
-  get parent() {
-    let parent = this.address.parent()
-    let identifier = parent.toString()
-    if (!identifier) return
-    while (!this.form.fields[identifier]) {
-      parent = parent.parent()
-      identifier = parent.toString()
-      if (!identifier) return
-    }
-    return this.form.fields[identifier]
-  }
-
-  get component() {
-    return [this.componentType, this.componentProps]
-  }
-
-  set component(value: FieldComponent<Component>) {
-    const component = toArr(value)
-    this.componentType = component[0]
-    this.componentProps = component[1] || {}
-  }
-
-  get decorator() {
-    return [this.decoratorType, this.decoratorProps]
-  }
-
-  set decorator(value: FieldDecorator<Decorator>) {
-    const decorator = toArr(value)
-    this.decoratorType = decorator[0]
-    this.decoratorProps = decorator[1] || {}
   }
 
   get selfErrors() {
@@ -396,69 +330,10 @@ export class Field<
     return this.form.getInitialValuesIn(this.path)
   }
 
-  get display(): FieldDisplayTypes {
-    const parentDisplay = this.parent?.display
-    if (parentDisplay && parentDisplay !== 'visible') {
-      if (this.selfDisplay && this.selfDisplay !== 'visible')
-        return this.selfDisplay
-      return parentDisplay
-    }
-    if (this.selfDisplay) return this.selfDisplay
-    return parentDisplay || this.form.display || 'visible'
-  }
-
-  get pattern(): FieldPatternTypes {
-    const parentPattern = this.parent?.pattern
-    if (this.selfPattern) return this.selfPattern
-    return parentPattern || this.form.pattern || 'editable'
-  }
-
   get required() {
     return parseValidatorDescriptions(this.validator).some(
       (desc) => desc.required
     )
-  }
-
-  get hidden() {
-    return this.display === 'hidden'
-  }
-
-  get visible() {
-    return this.display === 'visible'
-  }
-
-  set hidden(hidden: boolean) {
-    if (!isValid(hidden)) return
-    if (hidden) {
-      this.display = 'hidden'
-    } else {
-      this.display = 'visible'
-    }
-  }
-
-  set visible(visible: boolean) {
-    if (!isValid(visible)) return
-    if (visible) {
-      this.display = 'visible'
-    } else {
-      this.display = 'none'
-    }
-  }
-
-  get disabled() {
-    return this.pattern === 'disabled'
-  }
-
-  get readOnly() {
-    return this.pattern === 'readOnly'
-  }
-
-  get readPretty() {
-    return this.pattern === 'readPretty'
-  }
-
-  get editable() {
-    return this.pattern === 'editable'
   }
 
   get validateStatus() {
@@ -466,50 +341,6 @@ export class Field<
     if (this.selfInvalid) return 'error'
     if (this.selfWarnings.length) return 'warning'
     if (this.selfSuccesses.length) return 'success'
-  }
-
-  set readOnly(readOnly: boolean) {
-    if (!isValid(readOnly)) return
-    if (readOnly) {
-      this.pattern = 'readOnly'
-    } else {
-      this.pattern = 'editable'
-    }
-  }
-
-  set editable(editable: boolean) {
-    if (!isValid(editable)) return
-    if (editable) {
-      this.pattern = 'editable'
-    } else {
-      this.pattern = 'readPretty'
-    }
-  }
-
-  set disabled(disabled: boolean) {
-    if (!isValid(disabled)) return
-    if (disabled) {
-      this.pattern = 'disabled'
-    } else {
-      this.pattern = 'editable'
-    }
-  }
-
-  set readPretty(readPretty: boolean) {
-    if (!isValid(readPretty)) return
-    if (readPretty) {
-      this.pattern = 'readPretty'
-    } else {
-      this.pattern = 'editable'
-    }
-  }
-
-  set pattern(pattern: FieldPatternTypes) {
-    this.selfPattern = pattern
-  }
-
-  set display(display: FieldDisplayTypes) {
-    this.selfDisplay = display
   }
 
   set required(required: boolean) {
@@ -566,14 +397,6 @@ export class Field<
     })
   }
 
-  setTitle = (title?: TextType) => {
-    this.title = title
-  }
-
-  setDescription = (description?: TextType) => {
-    this.description = description
-  }
-
   setDataSource = (dataSource?: FieldDataSource) => {
     this.dataSource = dataSource
   }
@@ -614,14 +437,6 @@ export class Field<
     this.initialValue = initialValue
   }
 
-  setDisplay = (type?: FieldDisplayTypes) => {
-    this.display = type
-  }
-
-  setPattern = (type?: FieldPatternTypes) => {
-    this.pattern = type
-  }
-
   setLoading = (loading?: boolean) => {
     setLoading(this, loading)
   }
@@ -634,71 +449,9 @@ export class Field<
     setSubmitting(this, submitting)
   }
 
-  setComponent = <C extends JSXComponent, ComponentProps extends object = {}>(
-    component?: C,
-    props?: ComponentProps
-  ) => {
-    if (component) {
-      this.componentType = component as any
-    }
-    if (props) {
-      this.componentProps = this.componentProps || {}
-      Object.assign(this.componentProps, props)
-    }
-  }
-
-  setComponentProps = <ComponentProps extends object = {}>(
-    props?: ComponentProps
-  ) => {
-    if (props) {
-      this.componentProps = this.componentProps || {}
-      Object.assign(this.componentProps, props)
-    }
-  }
-
-  setDecorator = <D extends JSXComponent, ComponentProps extends object = {}>(
-    component?: D,
-    props?: ComponentProps
-  ) => {
-    if (component) {
-      this.decoratorType = component as any
-    }
-    if (props) {
-      this.decoratorProps = this.decoratorProps || {}
-      Object.assign(this.decoratorProps, props)
-    }
-  }
-
-  setDecoratorProps = <ComponentProps extends object = {}>(
-    props?: ComponentProps
-  ) => {
-    if (props) {
-      this.decoratorProps = this.decoratorProps || {}
-      Object.assign(this.decoratorProps, props)
-    }
-  }
-
   setState: IModelSetter<IFieldState> = createStateSetter(this)
 
   getState: IModelGetter<IFieldState> = createStateGetter(this)
-
-  onInit = () => {
-    this.initialized = true
-    initFieldUpdate(this)
-    this.notify(LifeCycleTypes.ON_FIELD_INIT)
-  }
-
-  onMount = () => {
-    this.mounted = true
-    this.unmounted = false
-    this.notify(LifeCycleTypes.ON_FIELD_MOUNT)
-  }
-
-  onUnmount = () => {
-    this.mounted = false
-    this.unmounted = true
-    this.notify(LifeCycleTypes.ON_FIELD_UNMOUNT)
-  }
 
   onInput = async (...args: any[]) => {
     if (args[0]?.target) {
@@ -747,35 +500,7 @@ export class Field<
     return batchReset(this, `${this.address}.**`, options)
   }
 
-  query = (pattern: FormPathPattern) => {
-    return new Query({
-      pattern,
-      base: this.address,
-      form: this.form,
-    })
-  }
-
   queryFeedbacks = (search?: ISearchFeedback): IFieldFeedback[] => {
     return queryFeedbacks(this, search)
-  }
-
-  notify = (type: LifeCycleTypes, payload?: any) => {
-    return this.form.notify(type, payload ?? this)
-  }
-
-  dispose = () => {
-    this.disposers.forEach((dispose) => {
-      dispose()
-    })
-    this.form.removeEffects(this)
-  }
-
-  destroy = () => {
-    this.dispose()
-    delete this.form.fields[this.address.toString()]
-  }
-
-  match = (pattern: FormPathPattern) => {
-    return FormPath.parse(pattern).matchAliasGroup(this.address, this.path)
   }
 }
